+++ lib/rubocop.rb
@@ -68,7 +68,7 @@ require 'rubocop/cop/word_array'
 require 'rubocop/cop/symbol_array'
 require 'rubocop/cop/trivial_accessors'
 require 'rubocop/cop/leading_comment_space'
-require 'rubocop/cop/semicolon_method_call'
+require 'rubocop/cop/colon_method_call'
 
 require 'rubocop/report/report'
 require 'rubocop/report/plain_text'