+++ spec/rubocop/cops/colon_method_call_spec.rb
@@ -4,8 +4,8 @@ require 'spec_helper'
 
 module Rubocop
   module Cop
-    describe SemicolonMethodCall do
-      let(:smc) { SemicolonMethodCall.new }
+    describe ColonMethodCall do
+      let(:smc) { ColonMethodCall.new }
 
       it 'registers an offence for instance method call' do
         inspect_source(smc,
@@ -14,10 +14,24 @@ module Rubocop
         expect(smc.offences.size).to eq(1)
       end
 
+      it 'registers an offence for instance method call with arg' do
+        inspect_source(smc,
+                       'file.rb',
+                       ['test::method_name(arg)'])
+        expect(smc.offences.size).to eq(1)
+      end
+
       it 'registers an offence for class method call' do
         inspect_source(smc,
                        'file.rb',
                        ['Class::method_name'])
+        expect(smc.offences.size).to eq(1)
+      end
+
+      it 'registers an offence for class method call with arg' do
+        inspect_source(smc,
+                       'file.rb',
+                       ['Class::method_name(arg, arg2)'])
         expect(smc.offences.size).to eq(1)
       end